tweak for #301, unnecessary regex?
It's not clear if both of these are necessary, the second one should
just match everything the first one matches plus things within internal
lines, so it seems redundant.

The unit test commited with it, plus all the others still pass, and
there's no diff in the output on a long test file.