Move back to call_user_func_array
This is equivalent, but removes the assumption that
parent constructor will also accept a single array
val as well as overloading. That assumption is true,
but may not always remain true.

Frankly, all of this is due for massive refactoring,
but that's something for... well, not today, that's
for sure!