Tighten legacy color conversion
The existing regex was only separating on comma,
but `var()` statements (which can be used as values)
can also include a comma (after which follows a
default value)

This seemed to be the case in one instance in Bootstrap,
causing it to be minified incorrectly.

This replacement is more strict, and will only
convert colors made up of their literal values,
ignoring any that include `var()`s in there.

Fixes #422